+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added methods `OrgVdcNetwork.Update`, `OrgVdcNetwork.UpdateAsync`, and `OrgVdcNetwork.Rename` [#292](https://github.com/vmware/go-vcloud-director/pull/292)
 * Added methods `EdgeGateway.Update` and `EdgeGateway.UpdateAsync` [#292](https://github.com/vmware/go-vcloud-director/pull/292)
+* Added methods  `catalog.QueryVappTemplateList`, `catalog.QueryCatalogItemList`, `catalog.QueryWithMetadataFields`, `catalog.QueryWithMetadataFilter` 
 
 ## 2.6.0 (March 13, 2010)